make countdown an effect
