+++ src/common/legacy.ts
@@ -1,4 +1,4 @@
-import { AddStreamTagEffect, ChatEffect, ChattersEffect, CommandEffectType, CommandVariableChange, DictLookupEffect, EmotesEffect, MadochanEffect, MediaEffect, RemoveStreamTagEffect, SetChannelGameIdEffect, SetChannelTitleEffect, VariableChangeEffect } from "../types"
+import { AddStreamTagEffect, ChatEffect, ChattersEffect, CommandEffectType, CommandVariableChange, CountdownAction, CountdownEffect, DictLookupEffect, EmotesEffect, MadochanEffect, MediaEffect, RemoveStreamTagEffect, SetChannelGameIdEffect, SetChannelTitleEffect, VariableChangeEffect } from "../types"
 
 const variableChangeToCommandEffect = (variableChange: CommandVariableChange): VariableChangeEffect => {
   return {
@@ -135,9 +135,31 @@ const chattersToCommandEffect = (_cmd: any): ChattersEffect => {
   }
 }
 
+const countdownToCommandEffect = (cmd: any): CountdownEffect => {
+  cmd.data.actions = (cmd.data.actions || []).map((action: CountdownAction) => {
+    if (typeof action.value === 'string') {
+      return action
+    }
+    if (action.value.sound && !action.value.sound.urlpath && action.value.sound.file) {
+      action.value.sound.urlpath = `/uploads/${encodeURIComponent(action.value.sound.file)}`
+    }
+
+    if (action.value.image && !action.value.image.urlpath && action.value.image.file) {
+      action.value.image.urlpath = `/uploads/${encodeURIComponent(action.value.image.file)}`
+    }
+    return action
+  })
+
+  return {
+    type: CommandEffectType.COUNTDOWN,
+    data: cmd.data
+  }
+}
+
 export default {
   addStreamTagsToCommandEffect,
   chattersToCommandEffect,
+  countdownToCommandEffect,
   setChannelTitleToCommandEffect,
   setChannelGameIdToCommandEffect,
   dictLookupToCommandEffect,
